Increasing Difficulty
Every 7 rooms explored increases the danger level by 1.
Unfortunately this also slows the game down a lot when there are a lot of enemies on the screen, but I figured it happens slowly and for now it is the easiest way to increase the difficulty slowly so the game doesn't go on forever (currently you could play forever and just get bored.